+++ dashboard/scans/forms.py
@@ -2,6 +2,7 @@ from flask_wtf import FlaskForm
 from wtforms import TextAreaField, SubmitField
 from wtforms.validators import DataRequired
 
+
 class ScanChecklistForm(FlaskForm):
     comment = TextAreaField(
         'Comment:',
