+++ dashboard/utils.py
@@ -1,10 +1,8 @@
-"""Location for utility functions (including those that might need datman)
+"""Location for utility functions (including those that need datman)
 """
 import os
 import json
 import time
-import glob
-import shutil
 import logging
 
 import datman.config
@@ -58,3 +56,29 @@ def get_software_version(json_contents):
     except KeyError:
         software_version = "Version Not Available"
     return software_name + " - " + software_version
+
+
+def update_header_diffs(scan):
+    site = scan.session.timepoint.site_id
+    config = datman.config.config(study=scan.get_study().id)
+
+    try:
+        tolerance = config.get_key("HeaderFieldTolerance", site=site)
+    except Exception:
+        tolerance = {}
+    try:
+        ignore = config.get_key("IgnoreHeaderFields", site=site)
+    except Exception:
+        ignore = []
+
+    tags = config.get_tags(site=site)
+    try:
+        qc_type = tags.get(scan.tag, "qc_type")
+    except KeyError:
+        check_bvals = False
+    else:
+        check_bvals = qc_type == 'dti'
+
+    scan.update_header_diffs(ignore=ignore,
+                             tolerance=tolerance,
+                             bvals=check_bvals)
